Changed from float to double

+++ src/model/game/Health.java
@@ -5,7 +5,7 @@ public interface Health {
 
     int getMax();
 
-    float getFraction();
+    double getFraction();
 
     boolean isDead();
